+++ lib/mjml/parser.rb
@@ -1,5 +1,3 @@
-require 'open3'
-
 module Mjml
   class Parser
     class ParseError < StandardError; end
@@ -33,11 +31,12 @@ module Mjml
     # Exec mjml command
     #
     # @return [String] The result as string
-    def run(in_tmp_file, beautify=true, minify=false, validation_level="soft")
+    def run(in_tmp_file, beautify=true, minify=false, validation_level="strict")
       Tempfile.create(["out", ".html"]) do |out_tmp_file|
         command = "-r #{in_tmp_file} -o #{out_tmp_file.path} --config.beautify #{beautify} --config.minify #{minify} --config.validationLevel #{validation_level}"
-        _, stderr, _ = Mjml.run_mjml(command)
-        raise ParseError.new(stderr.chomp) unless stderr.blank?
+        _, stderr, status = Mjml.run_mjml(command)
+        raise ParseError.new(stderr.chomp) unless status.success?
+        Mjml.logger.warn(stderr.chomp) unless stderr.blank?
         out_tmp_file.read
       end
     end